Separate cmake directory (#319)
* Split external deps to cmake dir

* Update clang-tidy config and fix two issues

* Update ide settings

* Remove compiler warnings

* Shifting to the new cmake/compilerwarnings.cmake
file brought many warnings. Many of these have been
suppressed here.

* Suppress double/float conversion warning

* Rename poke-in-the-eye hpp header

* Fix travis build (cmake bump)

* Silence warnings on gcc7 and clang

+++ src/aux/eigen_cerealisation.h
@@ -77,7 +77,7 @@ load(Archive &ar, Eigen::Matrix<_Scalar, _Rows, _Cols, _Options, _MaxRows, _MaxC
 
     matrix.resize(rows, cols);
 
-    ar(binary_data(matrix.data(), static_cast<std::size_t>(rows * cols * sizeof(_Scalar))));
+    ar(binary_data(matrix.data(), rows * cols * sizeof(_Scalar)));
 }
 
 } /* namespace cereal */
